feat(sdk): integrate with scroll-proving-sdk (#1)
Integrate with
[scroll-proving-sdk](https://github.com/scroll-tech/scroll-proving-sdk)
to support proof generation for Scroll SDK users.

The high level idea is that this binary will fetch proving tasks from
Scroll Coordinator and then send requests to the Snarkify platform for
proof generation. And it will check the job status periodically and
submit the results to Scroll Coordinator at the end.